requirements.txt
openpyxl==3.0.10
Pillow==9.4.0
